Se crearon los DTO de genero y el DTO que ya reune la locacion, la cantidad en total y por genero

+++ src/main/java/co/edu/umanizales/tads/model/Pet.java
@@ -3,7 +3,7 @@ package co.edu.umanizales.tads.model;
 public class Pet
 {
     private String name;
-    private int age;
+    private byte age;
     private String type;
     private String race;
     private char sex;
